+++ Sources/Restructure/Restructure.swift
@@ -3,7 +3,7 @@
 //  Restructure
 //
 //  Created by Stephen H. Gerstacker on 11/3/18.
-//  Copyright @ 2019 Stephen H. Gerstacker. All rights reserved.
+//  Copyright @ 2020 Stephen H. Gerstacker. All rights reserved.
 //
 
 import Foundation
@@ -66,6 +66,23 @@ public class Restructure {
     public internal(set) var userVersion: Int {
         get { get(pragma: "user_version") }
         set { set(pragma: "user_version", value: newValue) }
+    }
+
+    /// The underlying SQLite version {
+    public var sqliteVersion: String {
+        do {
+            let statement = try prepare(query: "SELECT sqlite_version()")
+            let result = statement.step()
+
+            switch result {
+            case let .row(row):
+                return row[0]
+            default:
+                fatalError("Failed to fetch sqlite_version from a result: \(result)")
+            }
+        } catch {
+            fatalError("Failed to fetch sqlite_version: \(error)")
+        }
     }